Module Eliom

module App : sig ... end

module Bus : sig ... end

Broadcasting facilities between clients and server

module Client : sig ... end

module Client_base : sig ... end

module Client_main : sig ... end

module Client_value : sig ... end

module Comet : sig ... end

Primitives to push data to the client, without explicit request.

module Comet_base : sig ... end

module Common : sig ... end

Low level functions for Eliom, exceptions and types.

module Common_base : sig ... end

module Config : sig ... end

This module makes also possible get information from the configuration file.

module Content : sig ... end

This module allows creating valid HTML content, or other XML formats.

module Content_core : sig ... end

See Content for complete module.

module Content_functor : sig ... end

module Content_sigs : sig ... end

module Cookies_base : sig ... end

module Cscache : sig ... end

Client-server cache for Eliom applications, that is used both for keeping data on client side or keeping values in memory during a request.

module Eliom_form : sig ... end

module Eliom_lazy : sig ... end

module Eliom_react : sig ... end

Propagate events occurrences from the server to the client and the other way around. Occurrence propagation is done asynchronously.

module Eliom_uri : sig ... end

Low-level functions for relative or absolute URL calculation.

module Error_pages : sig ... end

module Extension : sig ... end

Allows Ocsigen's extension to access Eliom data. See the Eliom manual for more information about Eliom's extensions

module Form_sigs : sig ... end

module Lib : sig ... end

A few common functions used by Eliom. Extension of OCaml stdlib. See also Ocsigen_base.Lib

module Lib_base : sig ... end

module Mkreg : sig ... end

This module defines the functor to use to creates modules generating functions to register services for your own types of pages. It is used for example in Registration.

module Mod_cli : sig ... end

module Mod_cookies : sig ... end

module Mod_datasess : sig ... end

module Mod_gc : sig ... end

module Mod_main : sig ... end

module Mod_pagegen : sig ... end

module Mod_parameters : sig ... end

module Mod_persess : sig ... end

module Mod_sersess : sig ... end

module Mod_sessadmin : sig ... end

module Mod_sessexpl : sig ... end

module Mod_sessiongroups : sig ... end

module Mod_timeouts : sig ... end

module Notif : sig ... end

Server to client notifications.

module Parameter : sig ... end

module Parameter_base : sig ... end

module Parameter_sigs : sig ... end

Ad-hoc runtime type representation for service parameters.

module Process : sig ... end

module Reference : sig ... end

module Registration : sig ... end

Eliom services registration for various kinds of page content: Eliom application, valid Html, actions, redirections, static files, ...

module Registration_sigs : sig ... end

module Request_info : sig ... end

This module contains the functions you need to get (or set) information about current request.

module Route : sig ... end

module Route_base : sig ... end

module Runtime : sig ... end

module Service : sig ... end

Creation and manipulation of Eliom services.

module Service_base : sig ... end

module Service_sigs : sig ... end

module Shared : sig ... end

This module implements shared (i.e., client-server) versions of the React and ReactiveData libraries.

module Shared_content : sig ... end

module Shared_sigs : sig ... end

module State : sig ... end

Storing server-side values for your applications or sessions.

module Syntax : sig ... end

module Tools : sig ... end

Helpers for (hierarchical) menu generation in HTML5. See the Eliom manual for more information about menu or hierarchical site.

module Types : sig ... end

Types shared by client and server.

module Types_base : sig ... end

module Wrap : sig ... end