Module Content.Svg

Building and pretty-printing valid SVG tree. Information about Svg api can be found at Svg_sigs.T

See the Eliom manual for more information ondom semantics vs. functional semantics for SVG tree manipulated by client/server application.

type +'a elt

type +'a attrib

type 'a wrap = 'a

type 'a list_wrap = 'a list

type uri = Xml.uri

module F : sig ... end

Typed interface for building valid SVG tree (functional semantics). See Svg_sigs.T.

module D : sig ... end

Typed interface for building valid SVG tree (DOM semantics). See Svg_sigs.T.

module R : sig ... end

Creation of SVG content from shared reactive signals and data (Shared). For the operations provided, see Svg_sigs.T.

module C : sig ... end

Creation of content from client-side values. This makes possible to insert in server side generated pages some nodes that will be computed on client side (for example reactive nodes).

module Id : sig ... end

Node identifiers.

module Printer : Xml_sigs.Typed_pp with type +'a elt := 'a elt and type doc := F.doc

SVG printer. See Xml_sigs.Typed_pp.