Module Mod_sessiongroups.Data

type group_of_group_data = [ `Session ] Common.sessgrp Ocsigen_base.Cache.Dlist.node

val add : ?set_max:int -> Common.sitedata -> string -> [< Common.cookie_level ] Common.sessgrp -> string Ocsigen_base.Cache.Dlist.node

val remove : 'a Ocsigen_base.Cache.Dlist.node -> unit

val remove_group : [< Common.cookie_level ] Common.sessgrp -> unit

val find : [< Common.cookie_level ] Common.sessgrp -> string Ocsigen_base.Cache.Dlist.t

returns the dlist containing all session group elements

val find_node_in_group_of_groups : [< `Session ] Common.sessgrp -> group_of_group_data option

Groups of browser sessions belong to a group of groups. As these groups are not associated to a cookie, we put this information here.

val move : ?set_max:int -> Common.sitedata -> string Ocsigen_base.Cache.Dlist.node -> [< Common.cookie_level ] Common.sessgrp -> string Ocsigen_base.Cache.Dlist.node

val up : string Ocsigen_base.Cache.Dlist.node -> unit

val nb_of_groups : unit -> int

val group_size : [< Common.cookie_level ] Common.sessgrp -> int

val set_max : 'a Ocsigen_base.Cache.Dlist.node -> int -> unit